+++ incubator/build.gradle.kts
@@ -1,0 +1,7 @@
+group = "org.traffichunter.titan.incubator"
+version = "1.0-SNAPSHOT"
+
+dependencies {
+    implementation(project(":bootstrap"))
+    implementation(project(":core"))
+}
