+++ settings.gradle.kts
@@ -11,3 +11,4 @@ include("titan-spring-client")
 include("smoke-test")
 include("smoke-test:smoke-spring")
 include("smoke-test:smoke-titan")
+include("incubator")
